Warn when audio samples are dropped (#1740)
Log a tracing::warn whenever audio samples are dropped in the macOS and Windows speaker code paths so we can observe and diagnose underruns/overflow events. In macOS, warn with the number of samples that couldn't be pushed into the producer buffer. In Windows, warn when the internal queue exceeds the 8192 sample cap and we drop the excess.